lambdas/certificates/requirements.txt
boto3==1.7.32
botocore==1.10.32
requests==2.18.4
urllib3==1.22
cryptography==2.8
kubernetes==10.0.1
ruamel.yaml==0.16.5
